gte_pcsx: Restore original type of BOUNDS func retval (PCSX Rearmed bug)
Fixes gfx glitch regression from one of the previous commits, titled
'gte_pcsx: Disable some MAC overflow checks on slow platforms (from PCSX Rearmed)'

This fixes glitched graphics in 1st level of 'Driver', that also occur
on PC builds of PCSX Rearmed. Will submit pull request there.

+++ src/gte/gte_pcsx/gte.cpp
@@ -174,7 +174,9 @@
 
 #define gteop (psxRegs.code & 0x1ffffff)
 
-INLINE s32 BOUNDS(s64 n_value, s64 n_max, int n_maxflag, s64 n_min, int n_minflag) {
+//senquack-Don't try to optimize return value to s32 like PCSX Rearmed did here:
+//it's why as of Nov. 2016, PC build has gfx glitches in 1st level of 'Driver'
+INLINE s64 BOUNDS(s64 n_value, s64 n_max, int n_maxflag, s64 n_min, int n_minflag) {
 	if (n_value > n_max) {
 		gteFLAG |= n_maxflag;
 	} else if (n_value < n_min) {
